refactor(web): extract useChatScroll composable from chat-pane
Move ~250 lines of scroll logic (stick-to-bottom follow, scroll-up
unlock, prepend position preservation, jump/highlight, cross-session
anchor isolation) out of the 2.5k-line chat-pane.vue into
composables/useChatScroll.ts. Behavior-preserving extraction with no
algorithm changes; carves a clean boundary for the upcoming
in-bottom threshold and spacer-pin work.
